requirements.txt
pymongo
ttkbootstrap
python-dotenv
Pillow